Identity and Authentication Implemented!
The website is now functional. The page All Messages was added for
convenience.
The Topics/Details page is still needing to have the UserName display
instead of the Id. This is a next fix.
Also to enhance, Edit/Delete links will be set only if the message and
User match.