Restructured Models and Views
After office hours with Brian Bird, I came to realize the unnecessary,
convoluted manner in which I was using Models. Therefore, I restructured
them along with the controllers and views that consumed them.
I attempted to fix the Migration issues I was having, but there is still
an error... The database now holds nothing. Seeding the database through
means of the dbInitializer does not work either...