Migration peculiarity and search need fix - Read description
Migration works well for creating users and their roles. I cannot figure
out how to place messages into the database though. I can get the topics
in, and I have tried many different methods to get the messages in and
attached to their topic.
When recreating the database, it is best to delete the mdf and ldf files
from App_Data and change the end number of the defaultconnection string
in the two places it is present (i.e. "EugeneCommunity#")

The search method for messages in both the messages and authorization
controller methods works, but it is failing to attach the member. This
still needs resolving.