Additions: add beer journal, view your beer journals. Delete beer journal is Almost working.
Two validation helper functions have been added - CheckFriendship and
ValidateAuthor.

Attempted to use the link anchors to jump to a beer journal instead of a
single journal entry page for viewing details. Not quite working yet,
however, the <a> tags and its correlated tag are written correctly.
Also, the ViewBag.ResultMessage isn't showing it's message...
Next to do is enhance these and make an edit page for a beer journal.